+++ NINA.Equipment/Equipment/MyFilterWheel/ASIFilterWheel.cs
@@ -28,27 +28,63 @@ using NINA.Equipment.Interfaces;
 namespace NINA.Equipment.Equipment.MyFilterWheel {
 
     public class ASIFilterWheel : BaseINPC, IFilterWheel {
-        private int id;
-        private IProfileService profileService;
+        private readonly int id;
+        private readonly IProfileService profileService;
+        private EFWdll.EFW_INFO efwInfo;
 
         public ASIFilterWheel(int idx, IProfileService profileService) {
+            EFWdll.EFW_ERROR_CODE rv;
+            this.profileService = profileService;
+
             _ = EFWdll.GetID(idx, out var id);
             this.id = id;
-            this.profileService = profileService;
+
+            // Keep trying to get properties if we are trying to connect to the filter wheel while it is initializing
+            do {
+                rv = EFWdll.GetProperty(id, out efwInfo);
+            } while (rv == EFWdll.EFW_ERROR_CODE.EFW_ERROR_MOVING);
+
+            if (string.IsNullOrEmpty(efwInfo.Name)) {
+                Logger.Error($"EFW: Unable to get filter wheel properties for EFW at index {idx}: {rv}");
+                return;
+            }
+
+            Name = efwInfo.Name;
+
+            CalibrateEfwCommand = new AsyncCommand<bool>(CalibrateEfw);
         }
 
-        public short InterfaceVersion => 1;
+        public int[] FocusOffsets => Filters.Select((x) => x.FocusOffset).ToArray();
 
-        public int[] FocusOffsets => this.Filters.Select((x) => x.FocusOffset).ToArray();
+        public string[] Names => Filters.Select((x) => x.Name).ToArray();
 
-        public string[] Names => this.Filters.Select((x) => x.Name).ToArray();
+        public bool Unidirectional {
+            get {
+                if (Connected) {
+                    _ = EFWdll.GetDirection(efwInfo.ID, out var unidirectional);
+                    return unidirectional;
+                }
+
+                return false;
+            }
+
+            set {
+                if (Connected) {
+                    Logger.Trace($"EFW: Setting Unidirectional to {value}");
+
+                    _ = EFWdll.SetDirection(efwInfo.ID, value);
+                    profileService.ActiveProfile.FilterWheelSettings.Unidirectional = value;
+                    RaisePropertyChanged();
+                }
+            }
+        }
 
         public IList<string> SupportedActions => new List<string>();
 
         public AsyncObservableCollection<FilterInfo> Filters {
             get {
                 var filtersList = profileService.ActiveProfile.FilterWheelSettings.FilterWheelFilters;
-                var positions = info.slotNum;
+                var positions = efwInfo.slotNum;
 
                 // Find duplicate positions due to data corruption and remove duplicates
                 var duplicates = filtersList.GroupBy(x => x.Position).Where(x => x.Count() > 1).ToList();
@@ -95,20 +131,15 @@ namespace NINA.Equipment.Equipment.MyFilterWheel {
             }
         }
 
-        public bool HasSetupDialog { get; } = false;
+        public bool HasSetupDialog => false;
 
-        public string Id {
-            get {
-                return $"{Name} {id}";
-            }
-        }
+        public string Id => string.IsNullOrEmpty(FilterWheelAlias) ? $"{Name} #{id}" : Name;
 
-        public string Name => "ZWOptical FilterWheel";
+        public string Name { get; private set; }
 
         public string Category => "ZWOptical";
 
         private bool _connected = false;
-        private EFWdll.EFW_INFO info;
 
         public bool Connected {
             get => _connected;
@@ -120,13 +151,13 @@ namespace NINA.Equipment.Equipment.MyFilterWheel {
 
         public string Description => "Native driver for ZWOptical filter wheels";
 
-        public string DriverInfo => "Native driver for ZWOptical filter wheels";
+        public string DriverInfo { get; private set; } = string.Empty;
 
         public string DriverVersion => "1.0";
 
         public short Position {
             get {
-                var err = EFWdll.GetPosition(this.info.ID, out var position);
+                var err = EFWdll.GetPosition(efwInfo.ID, out var position);
                 if (err == EFWdll.EFW_ERROR_CODE.EFW_SUCCESS) {
                     return (short)position;
                 } else {
@@ -136,7 +167,7 @@ namespace NINA.Equipment.Equipment.MyFilterWheel {
             }
 
             set {
-                var err = EFWdll.SetPosition(this.info.ID, value);
+                var err = EFWdll.SetPosition(efwInfo.ID, value);
                 if (err != EFWdll.EFW_ERROR_CODE.EFW_SUCCESS) {
                     Logger.Error($"EFW Communication error during position change {err}");
                 }
@@ -145,13 +176,21 @@ namespace NINA.Equipment.Equipment.MyFilterWheel {
 
         public Task<bool> Connect(CancellationToken token) {
             return Task.Run(() => {
-                if (EFWdll.Open(this.id) == EFWdll.EFW_ERROR_CODE.EFW_SUCCESS) {
-                    Connected = true;
+                EFWdll.EFW_ERROR_CODE rv;
 
-                    EFWdll.GetProperty(this.id, out var info);
-                    this.info = info;
+                if (EFWdll.Open(id) == EFWdll.EFW_ERROR_CODE.EFW_SUCCESS) {
+                    // Keep trying to get properties if we are trying to connect to the filter wheel while it is initializing
+                    do {
+                        rv = EFWdll.GetProperty(id, out efwInfo);
+                    } while (rv == EFWdll.EFW_ERROR_CODE.EFW_ERROR_MOVING);
 
-                    EFWdll.SetDirection(this.info.ID, false);
+                    if (rv != EFWdll.EFW_ERROR_CODE.EFW_SUCCESS) {
+                        Logger.Error($"EFW: Unable to get filter wheel properties for EFW at index {id}: {rv}");
+                        return false;
+                    }
+
+                    DriverInfo = $"SDK: {EFWdll.GetSDKVersion()}; FW: {GetFwVersionString()}";
+                    Unidirectional = profileService.ActiveProfile.FilterWheelSettings.Unidirectional;
 
                     Connected = true;
                     return true;
@@ -162,10 +201,109 @@ namespace NINA.Equipment.Equipment.MyFilterWheel {
             });
         }
 
+        // ZWO device alias is limited to 8 ASCII characters. Initialize with something longer to know we haven't yet asked the device for it
+        private string filterWheelAlias = "%%UNINITIALIZED%%";
+
+        public string FilterWheelAlias {
+            get {
+                if (filterWheelAlias.Equals("%%UNINITIALIZED%%")) {
+                    // We must connect to the filter wheel to get its ID. Quickly do this if we are not (such as during building the Chooser list)
+                    if (!Connected) {
+                        EFWdll.Open(efwInfo.ID);
+                    }
+
+                    filterWheelAlias = GetAlias();
+
+                    if (!Connected) {
+                        EFWdll.Close(efwInfo.ID);
+                    }
+
+                    Logger.Debug($"EFW: Filter wheel ID/Alias: {filterWheelAlias}");
+                }
+
+                return filterWheelAlias;
+            }
+
+            set {
+                if (!CanGetSetAlias) { return; }
+
+                Logger.Debug($"EFW: Setting Camera ID/Alias to: {value}");
+
+                EFWdll.SetID(efwInfo.ID, value);
+                filterWheelAlias = GetAlias();
+
+                _ = EFWdll.GetProperty(efwInfo.ID, out efwInfo);
+                Name = efwInfo.Name;
+
+                Logger.Info($"EFW: Filter wheel ID/Alias set to: {filterWheelAlias}");
+
+                RaisePropertyChanged();
+                RaisePropertyChanged(nameof(Name));
+                RaisePropertyChanged(nameof(Id));
+                profileService.ActiveProfile.FilterWheelSettings.Id = Id;
+            }
+        }
+
+        public bool CanGetSetAlias {
+            get {
+                _ = EFWdll.GetFirmwareVersion(efwInfo.ID, out var major, out var minor, out var patch);
+
+                if ((major * 100) + (minor * 10) + patch >= 309) {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+        }
+
+        // Come on, ZWO. Make some proper management interfaces
+        private string GetAlias() {
+
+            if (!CanGetSetAlias) { return string.Empty; }
+
+            _ = EFWdll.GetProperty(efwInfo.ID, out EFWdll.EFW_INFO info);
+
+            if (info.Name.Contains('(') && info.Name.Contains(')') && info.Name.EndsWith(")")) {
+                var openparen = info.Name.IndexOf('(');
+                var closeparen = info.Name.LastIndexOf(')');
+                var alias = closeparen - openparen;
+                return info.Name.Substring(openparen + 1, alias - 1);
+            } else {
+                return string.Empty;
+            }
+        }
+
+        private async Task<bool> CalibrateEfw(object arg) {
+            return await Task.Run(async () => {
+                EFWdll.EFW_ERROR_CODE rv;
+
+                if (Connected) {
+                    var currentPostion = Position;
+
+                    rv = EFWdll.Calibrate(efwInfo.ID);
+
+                    while (EFWdll.SetPosition(efwInfo.ID, currentPostion) == EFWdll.EFW_ERROR_CODE.EFW_ERROR_MOVING) {
+                        await Task.Delay(TimeSpan.FromSeconds(2));
+                    }
+
+                    return true;
+                } else {
+                    return false;
+                }
+            });
+        }
+
+        private string GetFwVersionString() {
+            _ = EFWdll.GetFirmwareVersion(efwInfo.ID, out var major, out var minor, out var patch);
+            return $"{major}.{minor}.{patch}";
+        }
+
         public void Disconnect() {
-            _ = EFWdll.Close(this.id);
+            _ = EFWdll.Close(id);
             this.Connected = false;
         }
+
+        public IAsyncCommand CalibrateEfwCommand { get; }
 
         public void SetupDialog() {
         }