+++ NINA.Equipment/SDK/CameraSDKs/ASISDK/ASICameraDll.cs
@@ -42,8 +42,6 @@ namespace ZWOptical.ASISDK {
             ASI_FAN_ON,
             ASI_PATTERN_ADJUST,
             ASI_ANTI_DEW_HEATER,
-            ASI_HUMIDITY,
-            ASI_ENABLE_DDR
         }
 
 
@@ -96,6 +94,7 @@ namespace ZWOptical.ASISDK {
             ASI_ERROR_VIDEO_MODE_ACTIVE,
             ASI_ERROR_EXPOSURE_IN_PROGRESS,
             ASI_ERROR_GENERAL_ERROR,//general error, eg: value is out of valid range
+            ASI_ERROR_INVALID_MODE,//the current mode is wrong
             ASI_ERROR_END
         };
         public enum ASI_BOOL {
@@ -132,12 +131,14 @@ namespace ZWOptical.ASISDK {
             public ASI_BOOL IsUSB3Host;
             public ASI_BOOL IsUSB3Camera;
             public float ElecPerADU;
+            public int BitDepth;
+            public ASI_BOOL IsTriggerCam;
 
-            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 24)]
-            public byte[] Unused;//[20];
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 16)]
+            public byte[] Unused;
 
             public string Name {
-                get { return Encoding.ASCII.GetString(name).TrimEnd((Char)0); }
+                get { return Encoding.ASCII.GetString(name).TrimEnd((char)0); }
             }
         };
 
@@ -167,10 +168,15 @@ namespace ZWOptical.ASISDK {
 
 
         public struct ASI_ID {
-            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 8)]
+            [MarshalAs(UnmanagedType.ByValArray, SizeConst = 8, ArraySubType = UnmanagedType.U1)]
             public byte[] id;
+
             public string ID {
-                get { return Encoding.ASCII.GetString(id).TrimEnd((Char)0); }
+                get {
+                    string idAscii = Encoding.ASCII.GetString(id);
+                    char[] trimChars = new char[1];
+                    return idAscii.TrimEnd(trimChars);
+                }
             }
         }
 
@@ -440,6 +446,23 @@ namespace ZWOptical.ASISDK {
 
             return version;
         }
+
+        [SecurityCritical]
+        public static string GetId(int cameraId) {
+            return ASIGetID(cameraId, out ASI_ID id) == ASI_ERROR_CODE.ASI_SUCCESS ? id.ID : string.Empty;
+        }
+
+        [SecurityCritical]
+        public static void SetId(int cameraId, string id) {
+            ASI_ID asiId = default;
+            asiId.id = new byte[8];
+
+            byte[] bytes = Encoding.Default.GetBytes(id);
+            bytes.CopyTo(asiId.id, 0);
+
+            CheckReturn(ASISetID(cameraId, asiId), MethodBase.GetCurrentMethod(), cameraId);
+        }
+
     }
 
     [Serializable]