+++ NINA.Profile/Interfaces/IFilterWheelSettings.cs
@@ -21,5 +21,6 @@ namespace NINA.Profile.Interfaces {
         ObserveAllCollection<FilterInfo> FilterWheelFilters { get; set; }
         string Id { get; set; }
         bool DisableGuidingOnFilterChange { get; set; }
+        bool Unidirectional { get; set; }
     }
 }